billing_market/requirements.txt
asgiref==3.7.2
Django==5.0.3
django-cors-headers==4.3.1
django-environ==0.11.2
django-phonenumber-field==7.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
mysqlclient==2.2.4
phonenumbers==8.13.31
PyJWT==2.8.0
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
